Move test samples to a separate directory
It is a nuisance to maintain the test samples with the pre-commit
checks. Therefore we move them to a separate directory, `test_samples/`
so that they are automatically excluded from the pre-commit checks.
