+++ test_samples/pyicontract_hypothesis/expected_ghostwrites/for_test_bare_and_explicit.txt
@@ -3,7 +3,7 @@ def test_some_func(self) -> None:
         x=integers(min_value=1)
     )
     def execute(kwargs) -> None:
-        tests.pyicontract_hypothesis.samples.sample_module.some_func(**kwargs)
+        test_samples.pyicontract_hypothesis.sample_module.some_func(**kwargs)
 
     execute()
 
@@ -12,7 +12,7 @@ def test_another_func(self) -> None:
         x=integers(min_value=1).filter(lambda x: square_greater_than_zero(x))
     )
     def execute(kwargs) -> None:
-        tests.pyicontract_hypothesis.samples.sample_module.another_func(**kwargs)
+        test_samples.pyicontract_hypothesis.sample_module.another_func(**kwargs)
 
     execute()
 
@@ -21,6 +21,6 @@ def test_yet_another_func(self) -> None:
         fixed_dictionaries({'x': integers(), 'y': integers()}).filter(lambda d: d['x'] < d['y'])
     )
     def execute(kwargs) -> None:
-        tests.pyicontract_hypothesis.samples.sample_module.yet_another_func(**kwargs)
+        test_samples.pyicontract_hypothesis.sample_module.yet_another_func(**kwargs)
 
     execute()
