+++ test_samples/pyicontract_hypothesis/expected_ghostwrites/for_test_non_bare_and_explicit.py
@@ -1,21 +1,21 @@
-"""Test tests.pyicontract_hypothesis.samples.sample_module with inferred Hypothesis strategies."""
+"""Test test_samples.pyicontract_hypothesis.sample_module with inferred Hypothesis strategies."""
 
 import unittest
 
 from hypothesis import given
 
-import tests.pyicontract_hypothesis.samples.sample_module
+import test_samples.pyicontract_hypothesis.sample_module
 
 
 class TestWithInferredStrategies(unittest.TestCase):
-    """Test all functions from tests.pyicontract_hypothesis.samples.sample_module with inferred Hypothesis strategies."""
+    """Test all functions from test_samples.pyicontract_hypothesis.sample_module with inferred Hypothesis strategies."""
 
     def test_some_func(self) -> None:
         @given(
             x=integers(min_value=1)
         )
         def execute(kwargs) -> None:
-            tests.pyicontract_hypothesis.samples.sample_module.some_func(**kwargs)
+            test_samples.pyicontract_hypothesis.sample_module.some_func(**kwargs)
 
         execute()
 
@@ -24,7 +24,7 @@ class TestWithInferredStrategies(unittest.TestCase):
             x=integers(min_value=1).filter(lambda x: square_greater_than_zero(x))
         )
         def execute(kwargs) -> None:
-            tests.pyicontract_hypothesis.samples.sample_module.another_func(**kwargs)
+            test_samples.pyicontract_hypothesis.sample_module.another_func(**kwargs)
 
         execute()
 
@@ -33,7 +33,7 @@ class TestWithInferredStrategies(unittest.TestCase):
             fixed_dictionaries({'x': integers(), 'y': integers()}).filter(lambda d: d['x'] < d['y'])
         )
         def execute(kwargs) -> None:
-            tests.pyicontract_hypothesis.samples.sample_module.yet_another_func(**kwargs)
+            test_samples.pyicontract_hypothesis.sample_module.yet_another_func(**kwargs)
 
         execute()
