+++ test_samples/pyicontract_hypothesis/expected_ghostwrites/for_test_path.txt
