+++ test_samples/pyicontract_hypothesis/well_formatted_with_two_arguments.py
